split portal types up, start implementing conversion
